+++ lib/ui/groups/GroupsHomePanel.dart
@@ -23,6 +23,7 @@ import 'package:illinois/ext/Group.dart';
 import 'package:illinois/model/Analytics.dart';
 import 'package:illinois/service/FlexUI.dart';
 import 'package:illinois/ui/attributes/ContentAttributesPanel.dart';
+import 'package:illinois/ui/profile/ProfileHomePanel.dart';
 import 'package:illinois/ui/widgets/RibbonButton.dart';
 import 'package:illinois/utils/AppUtils.dart';
 import 'package:rokwire_plugin/model/content_attributes.dart';
@@ -74,6 +75,9 @@ class _GroupsHomePanelState extends State<GroupsHomePanel> implements Notificati
   rokwire.GroupsContentType? _selectedContentType;
   bool _contentTypesVisible = false;
 
+  GestureRecognizer? _loginRecognizer;
+  GestureRecognizer? _selectAllRecognizer;
+
   List<Group>? _allGroups;
   List<Group>? _userGroups;
 
@@ -81,26 +85,29 @@ class _GroupsHomePanelState extends State<GroupsHomePanel> implements Notificati
 
   @override
   void initState() {
-    super.initState();
     NotificationService().subscribe(this, [
       Groups.notifyUserMembershipUpdated,
       Groups.notifyGroupCreated,
       Groups.notifyGroupUpdated,
       Groups.notifyGroupDeleted,
       Groups.notifyUserGroupsUpdated,
-      Auth2.notifyLoginSucceeded,
-      Auth2.notifyLogout,
+      Auth2.notifyLoginChanged,
       FlexUI.notifyChanged,
       Connectivity.notifyStatusChanged,
     ]);
+    _loginRecognizer = TapGestureRecognizer()..onTap = _onTapLogin;
+    _selectAllRecognizer = TapGestureRecognizer()..onTap = _onSelectAllGroups;
     _selectedContentType = widget.contentType;
     _reloadGroupsContent();
+    super.initState();
   }
 
   @override
   void dispose() {
-    super.dispose();
     NotificationService().unsubscribe(this);
+    _loginRecognizer?.dispose();
+    _selectAllRecognizer?.dispose();
+    super.dispose();
   }
 
   @override
@@ -158,7 +165,7 @@ class _GroupsHomePanelState extends State<GroupsHomePanel> implements Notificati
   }
 
   Future<List<Group>?> _loadUserGroups() async =>
-    Groups().loadGroups(contentType: rokwire.GroupsContentType.my);
+    Auth2().isLoggedIn ? Groups().loadGroups(contentType: rokwire.GroupsContentType.my) : null;
 
   Future<List<Group>?> _loadAllGroups() async =>
     Groups().loadGroups(
@@ -225,22 +232,26 @@ class _GroupsHomePanelState extends State<GroupsHomePanel> implements Notificati
     return
       Column(children: <Widget>[
         _buildGroupsContentSelection(),
-        Expanded(child: Stack(alignment: Alignment.topCenter, children: [
-          Column(children: [
-            _buildFunctionalBar(),
-            Expanded(child: _isLoading
-              ? Center(child: CircularProgressIndicator(strokeWidth: 2, valueColor: AlwaysStoppedAnimation<Color?>(Styles().colors.fillColorPrimary), ),)
-              : Container(color: Styles().colors.background, child:
-                  RefreshIndicator(onRefresh: _onPullToRefresh, child:
-                    SingleChildScrollView(scrollDirection: Axis.vertical, physics: AlwaysScrollableScrollPhysics(), child:
-                      Column(children: <Widget>[ _buildGroupsContent(), ],),
+        Expanded(child:
+          Stack(alignment: Alignment.topCenter, children: [
+            _isLoading
+              ? Center(child: CircularProgressIndicator(strokeWidth: 2, color: Styles().colors.fillColorPrimary),)
+              : Column(children: [
+                  if ((_selectedContentType != rokwire.GroupsContentType.my) || Auth2().isLoggedIn)
+                    _buildFunctionalBar(),
+                  Expanded(child:
+                    Container(color: Styles().colors.background, child:
+                      RefreshIndicator(onRefresh: _onPullToRefresh, child:
+                        SingleChildScrollView(scrollDirection: Axis.vertical, physics: AlwaysScrollableScrollPhysics(), child:
+                          Column(children: <Widget>[ _buildGroupsContent(), ],),
+                        ),
+                      ),
                     ),
                   ),
-                ),
-            )
-          ]),
-          _buildContentTypesContainer()
-        ]))
+                ]),
+            _buildContentTypesContainer()
+          ])
+        )
       ]);
   }
 
@@ -591,23 +602,29 @@ class _GroupsHomePanelState extends State<GroupsHomePanel> implements Notificati
 
   Widget _buildLoggedOutContent() {
     final String linkLoginMacro = "{{link.login}}";
-    String messageTemplate = Localization().getStringEx("panel.groups_home.label.my_groups.logged_out", "You are not logged in. To access your groups, you need to $linkLoginMacro first.");
+    String messageTemplate = Localization().getStringEx("panel.groups_home.label.my_groups.logged_out", "You are not logged in. To access your groups, $linkLoginMacro with your NetID and set your privacy level to 4 or 5 under Settings.");
     List<String> messages = messageTemplate.split(linkLoginMacro);
     List<InlineSpan> spanList = <InlineSpan>[];
     if (0 < messages.length)
       spanList.add(TextSpan(text: messages.first));
     for (int index = 1; index < messages.length; index++) {
-      spanList.add(TextSpan(text: Localization().getStringEx("panel.groups_home.label.my_groups.logged_out.link.login", "Login"), style : Styles().textStyles.getTextStyle("widget.link.button.title.regular"),
-        recognizer: TapGestureRecognizer()..onTap = _onTapLogin, ));
+      spanList.add(TextSpan(text: Localization().getStringEx("panel.groups_home.label.my_groups.logged_out.link.login", "sign in"), style : Styles().textStyles.getTextStyle("widget.link.button.title.regular"),
+        recognizer: _loginRecognizer, ));
       spanList.add(TextSpan(text: messages[index]));
     }
 
-    return Container(padding: EdgeInsets.symmetric(horizontal: 16, vertical: 30), child:
+    return Container(padding: EdgeInsets.symmetric(horizontal: 48, vertical: 32), child:
       RichText(textAlign: TextAlign.left, text:
         TextSpan(style: Styles().textStyles.getTextStyle("widget.message.dark.regular"), children: spanList)
       )
     );
   }
+
+  void _onTapLogin() {
+    Analytics().logSelect(target: "sign in");
+    ProfileHomePanel.present(context, content: ProfileContent.login,);
+  }
+
 
   Widget _buildEmptyMyGroupsContent() {
     return Container(padding: EdgeInsets.symmetric(horizontal: 16, vertical: 30), child:
@@ -615,7 +632,7 @@ class _GroupsHomePanelState extends State<GroupsHomePanel> implements Notificati
         TextSpan(style: Styles().textStyles.getTextStyle("widget.message.dark.regular"), children:[
           TextSpan(text:Localization().getStringEx("panel.groups_home.label.my_groups.empty", "You are not a member of any group. To join or create a group, see .")),
           TextSpan(text: Localization().getStringEx("panel.groups_home.label.my_groups.empty.link.all_groups", "All Groups"), style : Styles().textStyles.getTextStyle("widget.link.button.title.regular"),
-            recognizer: TapGestureRecognizer()..onTap = _onSelectAllGroups, ),
+            recognizer: _selectAllRecognizer, ),
           TextSpan(text:"."),
         ])
       )
@@ -718,16 +735,6 @@ class _GroupsHomePanelState extends State<GroupsHomePanel> implements Notificati
     }
   }
 
-  void _onTapLogin() {
-    Analytics().logSelect(target: "Login");
-    if (!FlexUI().isAuthenticationAvailable) {
-      AppAlert.showAuthenticationNAMessage(context);
-    }
-    else {
-      Auth2().authenticateWithOidc();
-    }
-  }
-
   void _syncAuthmanGroups() {
     Analytics().logSelect(target: "Sync Authman Group");
     Groups().syncAuthmanGroupsExt().then(
@@ -780,7 +787,7 @@ class _GroupsHomePanelState extends State<GroupsHomePanel> implements Notificati
     else if (name == Groups.notifyUserGroupsUpdated) {
       _applyUserGroups();
     }
-    else if ((name == Auth2.notifyLoginSucceeded) ||  (name == Auth2.notifyLogout)) {
+    else if (name == Auth2.notifyLoginChanged) {
       // Reload content with some delay, do not unmount immidately GroupsCard that could have updated the login state.
       Future.delayed(Duration(microseconds: 300), () {
         if (mounted) {